Display last exposure detection date
Why:
----
On the exposure history screen, the header is displaying the amount of
days on the calendar but not the time since the exposure history
detection process was run.

This PR:
----
- Add a module for exposing the last detection date from the native side
- Add component to display the time passed since the last exposure
detection happened

Reviewers:
----
From the debug menu, tap on the `Detect Exposures Now` button and then
when you navigate to the history screen, the time since the last
exposure detection was run should be a few seconds ago.

[trello](https://trello.com/c/mF6fMBd5)

Co-Authored-By: amandabeiner<aebeiner@gmail.com>
Co-Authored-By: Thomas Paresi<tparesi@gmail.com>